+++ lib/kcu/actions/set_secret_action.rb
@@ -1,0 +1,19 @@
+module Kcu
+  class SetSecretAction
+
+    extend LightService::Organizer
+
+    def self.call(resource, entry_name, entry_value)
+      ctx = with({
+        resource: resource,
+        entry_name: entry_name,
+        entry_value: entry_value,
+      }).reduce(
+        GetResourceNamespaceAndName,
+        Secrets::EncodeEntryValue,
+        Secrets::SetEntryValue,
+      )
+    end
+
+  end
+end